main.js: Apply relevant classes for system theme
This commit fixes a bug where the relevant class for the system theme,
i.e. `dark` or `light`, would not be applied when the system theme is
selected. This was not an issue in the original theme, but Tailwind
requires the theme to be explicitly stated in the HTML tree. (See
https://tailwindcss.com/docs/dark-mode#toggling-dark-mode-manually for
more details.)