Fixed the display issues
both types of display issues (smaller image in flex item in original web page; certain image shrinks more when window is smaller) are due to the fact that original image sources are relatively small.

So it can solved by setting the width and height in <img> to expand the image content in flex item.